[NNE] Replaced Logs with checks for detecting contract violations (precondition).
#jira UE-194213
#rb florent.guinier, nico.ranieri

#tests Compile Files (Non-Unity), Unreal Editor (Win64) Style-transfer and smoke tests

#rnx

[CL 30563302 by christian regg in ue5-main branch]

+++ Engine/Source/Runtime/NNE/Public/NNETypes.h
@@ -41,7 +41,7 @@ namespace UE::NNE
 	 * @param InType the type of the element to consider.
 	 * @return the data size in bytes of an element.
 	 */
-	size_t NNE_API GetTensorDataTypeSizeInBytes(ENNETensorDataType InType);
+	int32 NNE_API GetTensorDataTypeSizeInBytes(ENNETensorDataType InType);
 
 	/**
 	 * A symbolic tensor shape represents the shape of a tensor with potentially variable dimension.
